extract NeighborSimilarity to top level
